So:
I made a Drawable interface so that objects that need to be drawn like
markers and volume bars and stuff can be put into an arraylist and
redrawn if needed.  Only a draw(Graphics g) method is required to
satisfy the interface for maximum flexibility.

SettingsScreenController has a prototype volume bar now that responds to
clicking and dragging.  Volume can be retrieved by dividing the
rectangle width by 300.

Added Marker class, which is just a square drawable rectangle that is
always drawn centered.

Canvas changed to include List<Drawable> drawables, a list to track the
drawable objects on the screen, and addDrawable(Drawable d).